% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csib.r
\name{csib}
\alias{csib}
\title{Crossing SIBTEST}
\usage{
csib(
  data_ref,
  data_foc,
  minc = 2,
  cusr = 0.2,
  idw = 0,
  suspect_items,
  matching_items,
  listwise = 1
)
}
\arguments{
\item{data_ref}{The dataset that contains the reference group participants}

\item{data_foc}{The dataset that contains the focal group participants}

\item{minc}{The minimum number of individuals in a give total test score category. Initialized to 2 per Shealy and Stout (1993) recommendation}

\item{cusr}{The guess correction factor. This is initialized to 0.2, but can be changed based on the items used. The cusr value is the average of guess correction across the items in the dataset.}

\item{idw}{The weighting factor. This is initialized to use the reference and focal group weighting recommended by Shealy and Stout (1993)}

\item{suspect_items}{the item(s) to be assess for DIF. Listing one item number will assess for DIF on that item, multiple items (e.g. c(2,3,8)) will assess DIF for the bundle of items.}

\item{matching_items}{the list of items to match on.}

\item{listwise}{initialized to 1, this inidicates that the procedure will not delete any rows with missing data, changing this to 2 will listwise delete.}
}
\description{
This function assesses non-uniform DIF with one crossing point using the Crossing-SIBTEST procedure (Li & Stout, 1996).
The function outputs the non-uniform and uniform DIF magnitudes as well as the significance testing.
}
\examples{

\dontrun{

#perform Crossing-SIBTEST with one suspect item 

csib(data_ref = data_refC, data_foc = data_focC ,minc = 2, 
cusr = .2,  suspect_items = c(20), matching_items = c(1:19))

#perform Crossing-SIBTEST with one suspect item and no guessing (cusr = 0) 

csib(data_ref = data_refC, data_foc = data_focC, minc = 2, 
cusr = 0, suspect_items = c(20), matching_items = c(1:19))

#perform Crossing-SIBTEST with a bundle of suspect items 

csib(data_ref = data_refC, data_foc = data_focC, minc = 2, 
cusr = .2, suspect_items = c(16,17,18,19,20), matching_items = c(1:15))

}
}
\references{
Chalmers, R. P. (2018). Improving the Crossing-SIBTEST statistic for detecting non-uniform DIF. Psychometrika, 83, 2, 376-386.

DIF-Pack (2021) Measured Progress. https://psychometrics.onlinehelp.measuredprogress.org/tools/dif/

Li, H.-H. & Stout, W. (1996). A new procedure for detection of crossing DIF. Psychometrika, 61, 647-677.

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and detect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.
}
