% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_foc.R
\docType{data}
\name{data_foc}
\alias{data_foc}
\title{Example focal group data set for SIBTEST}
\format{
A data frame with 1000 rows and 30 variables:
}
\usage{
data_foc
}
\description{
A dataset containing simulated dichotomous response data
}
\keyword{datasets}
