% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_focC.R
\docType{data}
\name{data_focC}
\alias{data_focC}
\title{Example focal group data set for Crossing-SIBTEST}
\format{
A data frame with 1000 rows and 30 variables:
}
\usage{
data_focC
}
\description{
A dataset containing simulated dichotomous response data
}
\keyword{datasets}
