% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_focP.R
\docType{data}
\name{data_focP}
\alias{data_focP}
\title{Example focal group data set for POLYSIBTEST}
\format{
A data frame with 1000 rows and 20 variables:
}
\usage{
data_focP
}
\description{
A dataset containing simulated polytomous response data
}
\keyword{datasets}
