% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ref.R
\docType{data}
\name{data_ref}
\alias{data_ref}
\title{Example reference group data set for SIBTEST}
\format{
A data frame with 500 rows and 30 variables:
}
\usage{
data_ref
}
\description{
A dataset containing simulated dichotomous response data
}
\keyword{datasets}
