% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refC.R
\docType{data}
\name{data_refC}
\alias{data_refC}
\title{Example reference group data set for Crossing-SIBTEST}
\format{
A data frame with 500 rows and 30 variables:
}
\usage{
data_refC
}
\description{
A dataset containing simulated dichotomous response data
}
\keyword{datasets}
