% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refP.R
\docType{data}
\name{data_refP}
\alias{data_refP}
\title{Example reference group data set for POLYSIBTEST}
\format{
A data frame with 500 rows and 20 variables:
}
\usage{
data_refP
}
\description{
A dataset containing simulated polytomous response data
}
\keyword{datasets}
