% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psib.r
\name{psib}
\alias{psib}
\title{POLYSIBTEST}
\usage{
psib(
  data_ref,
  data_foc,
  minc = 2,
  cusr = 0,
  idw = 0,
  suspect_items,
  matching_items,
  listwise = 1,
  nch
)
}
\arguments{
\item{data_ref}{The dataset that contains the reference group participants}

\item{data_foc}{The dataset that contains the focal group participants}

\item{minc}{The minimum number of individuals in a give total test score category. Initialized to 2 per Shealy and Stout (1993) recommendation}

\item{cusr}{The guess correction factor. This is initialized to 0, but can be changed based on the items used. The cusr value is the average of guess correction across the items in the dataset.}

\item{idw}{The weighting factor. This is initialized to use the reference and focal group weighting recommended by Shealy and Stout (1993)}

\item{suspect_items}{the item(s) to be assess for DIF. Listing one item number will assess for DIF on that item, multiple items (e.g. c(2,3,8)) will assess DIF for the bundle of items.}

\item{matching_items}{the list of items to match on.}

\item{listwise}{initialized to 1, this inidicates that the procedure will not delete any rows with missing data, changing this to 2 will listwise delete.}

\item{nch}{a vector indicating the number of categories for each item. Example if there are 5 items and each has 3 catgories nch <- c(3,3,3,3,3). If there are 5 items and items 1 - 3 have 4 categories and items 4 and 5 have 2 then nch <- (4,4,4,2,2)}
}
\description{
This function assesses uniform DIF for polytomous DATA using the POLYSIBTEST procedure (Chang et al., 1996).
The function outputs the uniform DIF magnitudes as well as the significance testing.
}
\examples{

\dontrun{

#perform POLYSIBTEST with one suspect item. Sample data has 20 items with 5 categories each.
nch <- rep(5,20)
psib(data_ref = data_refP, data_foc = data_focP, 
minc = 2,suspect_items = c(20), matching_items = c(1:19),nch = nch)

#perform POLYSIBTEST with a bundle of suspect items 

psib(data_ref = data_refP, data_foc = data_focP, 
minc = 2, suspect_items = c(16,17,18,19,20), matching_items = c(1:15),nch = nch)

}
}
\references{
Chang, H. H., Mazzeo, J. & Roussos, L. (1996). DIF for Polytomously Scored Items: An Adaptation of the SIBTEST Procedure. Journal of Educational Measurement, 33, 333-353.

DIF-Pack (2021) Measured Progress. https://psychometrics.onlinehelp.measuredprogress.org/tools/dif/

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and detect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.
}
