% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib.r
\name{sib}
\alias{sib}
\title{SIBTEST}
\usage{
sib(
  data_ref,
  data_foc,
  minc = 2,
  cusr = 0.2,
  idw = 0,
  suspect_items,
  matching_items,
  listwise = 1
)
}
\arguments{
\item{data_ref}{The dataset that contains the reference group participants. Dichotomous data must only contain 0's and 1's.}

\item{data_foc}{The dataset that contains the focal group participants. Dichotomous data must only contain 0's and 1's.}

\item{minc}{The minimum number of individuals in a give total test score category. Initialized to 2 per Shealy and Stout (1993) recommendation}

\item{cusr}{The guess correction factor. This is initialized to 0.2, but can be changed based on the items used. The cusr value is the average of guess correction across the items in the dataset.}

\item{idw}{The weighting factor. This is initialized to use the reference and focal group weighting recommended by Shealy and Stout (1993)}

\item{suspect_items}{the item(s) to be assess for DIF. Listing one item number will assess for DIF on that item, multiple items (e.g. c(2,3,8)) will assess DIF for the bundle of items.}

\item{matching_items}{the list of items to match on.}

\item{listwise}{initialized to 1, this inidicates that the procedure will not delete any rows with missing data, changing this to 2 will listwise delete.}
}
\description{
This function assesses uniform DIF for dichotmous data using the SIBTEST procedure (Shealy & Stout, 1993) with a sophisticated regression correction (Jiang & Stout, 1998).
The function outputs the uniform DIF magnitudes as well as the significance testing.
}
\examples{

\dontrun{

#perform SIBTEST with one suspect item and the remaining items as a matching subtest

sib(data_ref = data_ref, data_foc = data_foc, minc = 2, 
cusr = .2, suspect_items = c(20), matching_items = c(1:19))

#perform SIBTEST with one suspect item and no guessing (cusr = 0) 

sib(data_ref = data_ref, data_foc = data_foc, minc = 2, 
cusr = 0, suspect_items = c(20), matching_items = c(1:19))

#perform SIBTEST with a bundle of suspect items and the remaining items as a 
matching subtest

sib(data_ref = data_ref, data_foc = data_foc, minc = 2, 
cusr = .2, suspect_items = c(16,17,18,19,20), matching_items = c(1:15))

}
}
\references{
DIF-Pack (2021) Measured Progress. https://psychometrics.onlinehelp.measuredprogress.org/tools/dif/

Jiang, H., & Stout, W. (1998). Improved Type I Error Control and Reduced Estimation Bias for DIF Detection Using SIBTEST. Journal of Educational and Behavioral Statistics, 23(4), 291–322. https://doi.org/10.3102/10769986023004291

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and detect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.

Weese, J. D., Turner, R. C., Ames, A., Crawford, B., & Liang, X. (2022). Reevaluating the SIBTEST Classification Heuristics for Dichotomous Differential Item Functioning. Educational and Psychological Measurement, 82(2), 307–329. https://doi.org/10.1177/00131644211017267

Weese, J. D. (2020). Development of an Effect Size to Classify the Magnitude of DIF in Dichotomous and Polytomous Items. Graduate Theses and Dissertations Retrieved from https://scholarworks.uark.edu/etd/3896
}
