% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBTEST_app.R
\name{SIBTEST_app}
\alias{SIBTEST_app}
\title{SIBTEST_app}
\usage{
SIBTEST_app()
}
\description{
This function provides a userinterface fo using SIBTEST, Crossing-SIBTEST, and POLYSIBTEST.
}
\examples{

\dontrun{

# Input files should be csv files
# When selecting the suspect items a single item can be typed as 21 or 23, but a bundle of items should be typed and separated by commas starting
# starting with the lowest item number to highest item number.
# Similarly the matching set of items should be typed and separated by commas starting with the lowest item number to highest item number.

SIBTEST_app()

}
}
\references{
Chalmers, R. P. (2018). Improving the Crossing-SIBTEST statistic for detecting non-uniform DIF. Psychometrika, 83, 2, 376-386.

Chang, H. H., Mazzeo, J. & Roussos, L. (1996). DIF for Polytomously Scored Items: An Adaptation of the SIBTEST Procedure. Journal of Educational Measurement, 33, 333-353.

DIF-Pack (2021) Measured Progress. https://psychometrics.onlinehelp.measuredprogress.org/tools/dif/

Jiang, H., & Stout, W. (1998). Improved Type I Error Control and Reduced Estimation Bias for DIF Detection Using SIBTEST. Journal of Educational and Behavioral Statistics, 23(4), 291–322. https://doi.org/10.3102/10769986023004291

Li, H.-H. & Stout, W. (1996). A new procedure for detection of crossing DIF. Psychometrika, 61, 647-677.

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and detect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.
}
